--- client/src/pages/Home.js
@@ -1,9 +1,0 @@
-const Home = () => {
-    return (
-        <aside>
-
-        </aside>
-    )
-}
-
-export default Home;
